clear code
