+++ novawallet/Modules/WatchOnly/CreateWatchOnly/CreateWatchOnlyViewController.swift
@@ -159,10 +159,10 @@ private extension CreateWatchOnlyViewController {
 
     func updateActionButtonState() {
         guard rootView.walletNameInputView.completed else {
-            setDisabledButton { $0.createWatchOnlyMissingNickname() }
+            setDisabledButton { $0.createWatchOnlyMissingWalletName() }
             return
         }
-        guard rootView.walletNameInputView.completed else {
+        guard rootView.substrateAddressInputView.completed else {
             setDisabledButton { $0.createWatchOnlyMissingSubstrate() }
             return
         }
@@ -274,44 +274,6 @@ private extension CreateWatchOnlyViewController {
 
     @objc func actionTermsTap() {
         presenter.toggleTermsCheckbox()
-    }
-}
-
-// MARK: - KeyboardAdoptable
-
-extension CreateWatchOnlyViewController: KeyboardAdoptable {
-    func updateWhileKeyboardFrameChanging(_ frame: CGRect) {
-        let localKeyboardFrame = view.convert(frame, from: nil)
-        let bottomInset = view.bounds.height - localKeyboardFrame.minY
-        let scrollView = rootView.containerView.scrollView
-        let scrollViewOffset = view.bounds.height - scrollView.frame.maxY
-
-        var contentInsets = scrollView.contentInset
-        contentInsets.bottom = max(0.0, bottomInset - scrollViewOffset)
-        scrollView.contentInset = contentInsets
-
-        if contentInsets.bottom > 0.0 {
-            let targetView: UIView?
-
-            if rootView.walletNameInputView.textField.isFirstResponder {
-                targetView = rootView.walletNameInputView
-            } else if rootView.substrateAddressInputView.textField.isFirstResponder {
-                targetView = rootView.substrateAddressInputView
-            } else if rootView.evmAddressInputView.textField.isFirstResponder {
-                targetView = rootView.evmAddressInputView
-            } else {
-                targetView = nil
-            }
-
-            if let firstResponderView = targetView {
-                let fieldFrame = scrollView.convert(
-                    firstResponderView.frame,
-                    from: firstResponderView.superview
-                )
-
-                scrollView.scrollRectToVisible(fieldFrame, animated: true)
-            }
-        }
     }
 }
